Move DEFAULT_DATA_PICKER_FILTERS to constants
